Updates LICENSE and NOTICE to comply with section 4d of the Apache License Version 2.0
Modifications are based on the instructions found at https://infra.apache.org/licensing-howto.html and using Apache Spark as an example.

Summary of changes:
* Updated LICENSE, NOTICE and pom.xml
* Removed code copied from Apache Commons Lang
* Updated code copied from Apache Commons Validator, removed all modifications and moved into a separate, shaded  package
* Refactored format validators to used the updated, unmodified code from Apache Commons Validator

Resolves #551
