+++ src/main/java/com/networknt/org/apache/commons/validator/routines/InetAddressValidator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.networknt.schema.format;
+package com.networknt.org.apache.commons.validator.routines;
 
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -31,10 +31,11 @@ import java.util.List;
  * This class is a Singleton; you can retrieve the instance via the {@link #getInstance()} method.
  * </p>
  *
- * @version $Revision$
- * @since Validator 1.4
+ * @since 1.4
  */
 public class InetAddressValidator implements Serializable {
+
+    private static final int MAX_BYTE = 128;
 
     private static final int IPV4_MAX_OCTET_VALUE = 255;
 
@@ -58,14 +59,11 @@ public class InetAddressValidator implements Serializable {
      */
     private static final InetAddressValidator VALIDATOR = new InetAddressValidator();
 
-    /**
-     * IPv4 RegexValidator
-     */
+    /** IPv4 RegexValidator */
     private final RegexValidator ipv4Validator = new RegexValidator(IPV4_REGEX);
 
     /**
      * Returns the singleton instance of this validator.
-     *
      * @return the singleton instance of this validator
      */
     public static InetAddressValidator getInstance() {
@@ -73,32 +71,30 @@ public class InetAddressValidator implements Serializable {
     }
 
     /**
-     * Checks if the specified string is a valid IP address.
-     *
+     * Checks if the specified string is a valid IPv4 or IPv6 address.
      * @param inetAddress the string to validate
      * @return true if the string validates as an IP address
      */
-    public boolean isValid(String inetAddress) {
+    public boolean isValid(final String inetAddress) {
         return isValidInet4Address(inetAddress) || isValidInet6Address(inetAddress);
     }
 
     /**
      * Validates an IPv4 address. Returns true if valid.
-     *
      * @param inet4Address the IPv4 address to validate
      * @return true if the argument contains a valid IPv4 address
      */
-    public boolean isValidInet4Address(String inet4Address) {
+    public boolean isValidInet4Address(final String inet4Address) {
         // verify that address conforms to generic IPv4 format
-        String[] groups = ipv4Validator.match(inet4Address);
+        final String[] groups = ipv4Validator.match(inet4Address);
 
         if (groups == null) {
             return false;
         }
 
         // verify that address subgroups are legal
-        for (String ipSegment : groups) {
-            if (ipSegment == null || ipSegment.length() == 0) {
+        for (final String ipSegment : groups) {
+            if (ipSegment == null || ipSegment.isEmpty()) {
                 return false;
             }
 
@@ -106,7 +102,7 @@ public class InetAddressValidator implements Serializable {
 
             try {
                 iIpSegment = Integer.parseInt(ipSegment);
-            } catch (NumberFormatException e) {
+            } catch (final NumberFormatException e) {
                 return false;
             }
 
@@ -125,13 +121,39 @@ public class InetAddressValidator implements Serializable {
 
     /**
      * Validates an IPv6 address. Returns true if valid.
-     *
      * @param inet6Address the IPv6 address to validate
      * @return true if the argument contains a valid IPv6 address
+     *
      * @since 1.4.1
      */
     public boolean isValidInet6Address(String inet6Address) {
-        boolean containsCompressedZeroes = inet6Address.contains("::");
+        String[] parts;
+        // remove prefix size. This will appear after the zone id (if any)
+        parts = inet6Address.split("/", -1);
+        if (parts.length > 2) {
+            return false; // can only have one prefix specifier
+        }
+        if (parts.length == 2) {
+            if (!parts[1].matches("\\d{1,3}")) {
+                return false; // not a valid number
+            }
+            final int bits = Integer.parseInt(parts[1]); // cannot fail because of RE check
+            if (bits < 0 || bits > MAX_BYTE) {
+                return false; // out of range
+            }
+        }
+        // remove zone-id
+        parts = parts[0].split("%", -1);
+        if (parts.length > 2) {
+            return false;
+        }
+        // The id syntax is implementation independent, but it presumably cannot allow:
+        // whitespace, '/' or '%'
+        if ((parts.length == 2) && !parts[1].matches("[^\\s/%]+")) {
+            return false; // invalid id
+        }
+        inet6Address = parts[0];
+        final boolean containsCompressedZeroes = inet6Address.contains("::");
         if (containsCompressedZeroes && (inet6Address.indexOf("::") != inet6Address.lastIndexOf("::"))) {
             return false;
         }
@@ -141,14 +163,14 @@ public class InetAddressValidator implements Serializable {
         }
         String[] octets = inet6Address.split(":");
         if (containsCompressedZeroes) {
-            List<String> octetList = new ArrayList<String>(Arrays.asList(octets));
+            final List<String> octetList = new ArrayList<>(Arrays.asList(octets));
             if (inet6Address.endsWith("::")) {
                 // String.split() drops ending empty segments
                 octetList.add("");
             } else if (inet6Address.startsWith("::") && !octetList.isEmpty()) {
                 octetList.remove(0);
             }
-            octets = octetList.toArray(new String[octetList.size()]);
+            octets = octetList.toArray(new String[0]);
         }
         if (octets.length > IPV6_MAX_HEX_GROUPS) {
             return false;
@@ -156,8 +178,8 @@ public class InetAddressValidator implements Serializable {
         int validOctets = 0;
         int emptyOctets = 0; // consecutive empty chunks
         for (int index = 0; index < octets.length; index++) {
-            String octet = octets[index];
-            if (octet.length() == 0) {
+            final String octet = octets[index];
+            if (octet.isEmpty()) {
                 emptyOctets++;
                 if (emptyOctets > 1) {
                     return false;
@@ -178,7 +200,7 @@ public class InetAddressValidator implements Serializable {
                 int octetInt = 0;
                 try {
                     octetInt = Integer.parseInt(octet, BASE_16);
-                } catch (NumberFormatException e) {
+                } catch (final NumberFormatException e) {
                     return false;
                 }
                 if (octetInt < 0 || octetInt > MAX_UNSIGNED_SHORT) {
