+++ src/main/java/com/networknt/schema/FormatKeyword.java
@@ -17,18 +17,15 @@
 package com.networknt.schema;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.networknt.schema.format.DurationValidator;
-import com.networknt.schema.format.EmailValidator;
+import com.networknt.schema.format.DateTimeValidator;
+import com.networknt.schema.format.DurationFormat;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
 
 public class FormatKeyword implements Keyword {
-    private static final String DATE = "date";
     private static final String DATE_TIME = "date-time";
-    private static final String UUID = "uuid";
-    private static final String EMAIL = "email";
     private static final String DURATION = "duration";
 
     private final ValidatorTypeCode type;
@@ -40,48 +37,42 @@ public class FormatKeyword implements Keyword {
     }
 
     Collection<Format> getFormats() {
-        return Collections.unmodifiableCollection(formats.values());
+        return Collections.unmodifiableCollection(this.formats.values());
     }
 
     @Override
-    public JsonValidator newValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext)
-            throws Exception {
+    public JsonValidator newValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext) {
         Format format = null;
         if (schemaNode != null && schemaNode.isTextual()) {
             String formatName = schemaNode.textValue();
-            format = formats.get(formatName);
-            // if you set custom format, override default Email/DateTime/UUID Validator
-            if (format != null) {
-                return new FormatValidator(schemaPath, schemaNode, parentSchema, validationContext, format, type);
-            }
-            // Validate date and time separately
-            if (formatName.equals(DATE) || formatName.equals(DATE_TIME)) {
-                ValidatorTypeCode typeCode = ValidatorTypeCode.DATETIME;
-                // Set custom error message
-                typeCode.setCustomMessage(type.getCustomMessage());
-                return new DateTimeValidator(schemaPath, schemaNode, parentSchema, validationContext, formatName, typeCode);
-            } else if (formatName.equals(UUID)) {
-                ValidatorTypeCode typeCode = ValidatorTypeCode.UUID;
-                // Set custom error message
-                typeCode.setCustomMessage(type.getCustomMessage());
-                return new UUIDValidator(schemaPath, schemaNode, parentSchema, validationContext, formatName, typeCode);
-            } else if (formatName.equals(EMAIL)) {
-                return new EmailValidator(schemaPath, schemaNode, parentSchema, validationContext, formatName, type);
-            }
-            else if (formatName.equals(DURATION)) {
-                return new DurationValidator(schemaPath, schemaNode, parentSchema, validationContext, formatName, type);
+            switch (formatName) {
+                case DURATION:
+                    format = new DurationFormat(validationContext.getConfig().isStrict(DURATION));
+                    break;
+
+                case DATE_TIME: {
+                    ValidatorTypeCode typeCode = ValidatorTypeCode.DATETIME;
+                    // Set custom error message
+                    typeCode.setCustomMessage(this.type.getCustomMessage());
+                    return new DateTimeValidator(schemaPath, schemaNode, parentSchema, validationContext, typeCode);
+                }
+
+                default:
+                    format = this.formats.get(formatName);
+                    break;
             }
         }
-        return new FormatValidator(schemaPath, schemaNode, parentSchema, validationContext, format, type);
+
+        return new FormatValidator(schemaPath, schemaNode, parentSchema, validationContext, format, this.type);
     }
 
     @Override
     public String getValue() {
-        return type.getValue();
+        return this.type.getValue();
     }
 
     @Override
     public void setCustomMessage(String message) {
-        type.setCustomMessage(message);
+        this.type.setCustomMessage(message);
     }
 }
