+++ src/main/java/com/networknt/schema/format/DateTimeValidator.java
@@ -14,69 +14,66 @@
  * limitations under the License.
  */
 
-package com.networknt.schema;
+package com.networknt.schema.format;
 
 import com.ethlo.time.ITU;
 import com.ethlo.time.LeapSecondException;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.networknt.schema.BaseJsonValidator;
+import com.networknt.schema.JsonSchema;
+import com.networknt.schema.JsonType;
+import com.networknt.schema.TypeFactory;
+import com.networknt.schema.ValidationContext;
+import com.networknt.schema.ValidationMessage;
+import com.networknt.schema.ValidatorTypeCode;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.time.LocalDate;
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-public class DateTimeValidator extends BaseJsonValidator implements JsonValidator {
+public class DateTimeValidator extends BaseJsonValidator {
     private static final Logger logger = LoggerFactory.getLogger(DateTimeValidator.class);
-    private static final String DATE = "date";
     private static final String DATETIME = "date-time";
 
-    private final String formatName;
-
-    public DateTimeValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext, String formatName, ValidatorTypeCode type) {
+    public DateTimeValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext, ValidatorTypeCode type) {
         super(schemaPath, schemaNode, parentSchema, type, validationContext);
-        this.formatName = formatName;
+
         this.validationContext = validationContext;
         parseErrorCode(getValidatorType().getErrorCodeKey());
-
-        if (!formatName.equals(DATE) && !formatName.equals(DATETIME)) {
-            throw new IllegalArgumentException(String.format("formatName must be one of the following: [%s, %s]", DATE, DATETIME));
-        }
     }
 
+    @Override
     public Set<ValidationMessage> validate(JsonNode node, JsonNode rootNode, String at) {
         debug(logger, node, rootNode, at);
 
-        Set<ValidationMessage> errors = new LinkedHashSet<ValidationMessage>();
+        Set<ValidationMessage> errors = new LinkedHashSet<>();
 
         JsonType nodeType = TypeFactory.getValueNodeType(node, this.validationContext.getConfig());
         if (nodeType != JsonType.STRING) {
             return errors;
         }
         if (!isLegalDateTime(node.textValue())) {
-            errors.add(buildValidationMessage(at, node.textValue(), formatName));
+            errors.add(buildValidationMessage(at, node.textValue(), DATETIME));
         }
         return Collections.unmodifiableSet(errors);
     }
 
-    private boolean isLegalDateTime(String string) {
+    private static boolean isLegalDateTime(String string) {
         try {
-            if (formatName.equals(DATE)) {
-                LocalDate.parse(string);
-            } else if (formatName.equals(DATETIME)) {
-                try {
-                    ITU.parseDateTime(string);
-                } catch (LeapSecondException ex) {
-                    if (!ex.isVerifiedValidLeapYearMonth()) {
-                        return false;
-                    }
+            try {
+                ITU.parseDateTime(string);
+            } catch (LeapSecondException ex) {
+                if (!ex.isVerifiedValidLeapYearMonth()) {
+                    return false;
                 }
             }
 
             return true;
         } catch (Exception ex) {
-            logger.debug("Invalid {}: {}", formatName, ex.getMessage());
+            logger.debug("Invalid {}: {}", DATETIME, ex.getMessage());
             return false;
         }
     }
