--- src/main/java/com/networknt/schema/format/EmailValidator.java
@@ -1,167 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.networknt.schema.format;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.networknt.schema.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Collections;
-import java.util.LinkedHashSet;
-import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-/**
- * <p>Perform email validations.</p>
- * <p>
- * Based on a script by <a href="mailto:stamhankar@hotmail.com">Sandeep V. Tamhankar</a>
- * http://javascript.internet.com
- * </p>
- * <p>
- * This implementation is not guaranteed to catch all possible errors in an email address.
- * </p>.
- *
- * @version $Revision$
- * @since Validator 1.4
- */
-public class EmailValidator extends BaseJsonValidator {
-    private static final Logger logger = LoggerFactory.getLogger(EmailValidator.class);
-
-    private static final String SPECIAL_CHARS = "\\p{Cntrl}\\(\\)<>@,;:'\\\\\\\"\\.\\[\\]";
-    private static final String VALID_CHARS = "(\\\\.)|[^\\s" + SPECIAL_CHARS + "]";
-    private static final String QUOTED_USER = "(\"(\\\\\"|[^\"])*\")";
-    private static final String WORD = "((" + VALID_CHARS + "|')+|" + QUOTED_USER + ")";
-
-    private static final String EMAIL_REGEX = "^\\s*?(.+)@(.+?)\\s*$";
-    private static final String IP_DOMAIN_REGEX = "^\\[(.*)\\]$";
-    private static final String USER_REGEX = "^\\s*" + WORD + "(\\." + WORD + ")*$";
-
-    private static final Pattern EMAIL_PATTERN = Pattern.compile(EMAIL_REGEX);
-    private static final Pattern IP_DOMAIN_PATTERN = Pattern.compile(IP_DOMAIN_REGEX);
-    private static final Pattern USER_PATTERN = Pattern.compile(USER_REGEX);
-
-    private static final int MAX_USERNAME_LEN = 64;
-
-    private static final boolean ALLOW_LOCAL = false;
-    private static final boolean ALLOW_TLD = false;
-
-    private final String formatName;
-
-
-
-    /**
-     * <p>Checks if a field has a valid e-mail address.</p>
-     *
-     * @param email The value validation is being performed on.  A <code>null</code>
-     *              value is considered invalid.
-     * @return true if the email address is valid.
-     */
-    public boolean isValid(String email) {
-        if (email == null) {
-            return false;
-        }
-
-        if (email.endsWith(".")) { // check this first - it's cheap!
-            return false;
-        }
-
-        // Check the whole email address structure
-        Matcher emailMatcher = EMAIL_PATTERN.matcher(email);
-        if (!emailMatcher.matches()) {
-            return false;
-        }
-
-        if (!isValidUser(emailMatcher.group(1))) {
-            return false;
-        }
-
-        if (!isValidDomain(emailMatcher.group(2))) {
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Returns true if the domain component of an email address is valid.
-     *
-     * @param domain being validated, may be in IDN format
-     * @return true if the email address's domain is valid.
-     */
-    protected boolean isValidDomain(String domain) {
-        // see if domain is an IP address in brackets
-        Matcher ipDomainMatcher = IP_DOMAIN_PATTERN.matcher(domain);
-        if (ipDomainMatcher.matches()) {
-            String ipAddress = ipDomainMatcher.group(1);
-            if (ipAddress.startsWith("IPv6:")) {
-                ipAddress = ipAddress.substring("IPv6:".length());
-                return InetAddressValidator.getInstance().isValidInet6Address(ipAddress);
-            }
-
-            return InetAddressValidator.getInstance().isValidInet4Address(ipAddress);
-        }
-
-        // Domain is symbolic name
-        DomainValidator domainValidator = DomainValidator.getInstance(ALLOW_LOCAL);
-        if (ALLOW_TLD) {
-            return domainValidator.isValid(domain) || (!domain.startsWith(".") && domainValidator.isValidTld(domain));
-        } else {
-            return domainValidator.isValid(domain);
-        }
-    }
-
-    /**
-     * Returns true if the user component of an email address is valid.
-     *
-     * @param user being validated
-     * @return true if the user name is valid.
-     */
-    protected boolean isValidUser(String user) {
-
-        if (user == null || user.length() > MAX_USERNAME_LEN) {
-            return false;
-        }
-
-        return USER_PATTERN.matcher(user).matches();
-    }
-
-    public EmailValidator(String schemaPath, JsonNode schemaNode, JsonSchema parentSchema, ValidationContext validationContext, String formatName,  ValidatorTypeCode type) {
-        super(schemaPath, schemaNode, parentSchema, type, validationContext);
-        this.formatName = formatName;
-        this.validationContext = validationContext;
-        parseErrorCode(getValidatorType().getErrorCodeKey());
-    }
-
-    @Override
-    public Set<ValidationMessage> validate(JsonNode node, JsonNode rootNode, String at) {
-        debug(logger, node, rootNode, at);
-
-        Set<ValidationMessage> errors = new LinkedHashSet<ValidationMessage>();
-
-        JsonType nodeType = TypeFactory.getValueNodeType(node, this.validationContext.getConfig());
-        if (nodeType != JsonType.STRING) {
-            return errors;
-        }
-        if (!isValid(node.textValue())) {
-            errors.add(buildValidationMessage(at, node.textValue(), formatName));
-        }
-        return Collections.unmodifiableSet(errors);
-
-    }
-}
